15. 3D Scroll Animation Completed
GeometryReader
The GeometryReader is great for detecting the position and size of your view. Let's wrap this around the CourseView component and set a frame. We're getting a geometry value that automatically listens to changes.

GeometryReader { geometry in
    SectionView(section: item)
}
.frame(width: 275, height: 275)
Get X Scroll Position
To get the X position, we can use the minX value from the frame. Think of the geometry as a box where minX is the starting left position and maxX as the ending right position.

geometry.frame(in: .global).minX
Apply Transformation
The minX position will change as you scroll. Using this dynamic value, we'll apply to the degrees of a 3D rotation effect. Note that it's important to convert our value to a Double. Also, since the X value changes too rapidly, we can divide that by 20. Finally, we'll need an offset of 40 because of our padding.

.rotation3DEffect(Angle(degrees:
        Double(geometry.frame(in: .global).minX - 30) / -20
    ), axis: (x: 0, y: 10.0, z: 0))